Move object path to its own directory
